update code thumbnail_iamge
