+++ BE/src/main/java/com/example/demo/Dto/ProductDTO/ProductDTO.java
@@ -26,8 +26,8 @@ public class ProductDTO implements Serializable {
     @DecimalMin(value = "0.0", inclusive = false, message = "PRICE_INVALID") // Giá trị tối thiểu lớn hơn 0
     private BigDecimal price;
     private int status_id;
-   @NotNull(message = "MUST_REQUIRED")
-   private String images;
+//   @NotNull(message = "MUST_REQUIRED")
+//   private String images;
     @NotNull(message = "MUST_REQUIRED")
     private int category_id;
     @NotNull(message = "MUST_REQUIRED")
