+++ BE/src/main/java/com/example/demo/Dto/ProductDTO/Product_Thumbnail.java
@@ -1,0 +1,12 @@
+package com.example.demo.Dto.ProductDTO;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Product_Thumbnail {
+    private String fullPath;
+}
