+++ BE/src/main/java/com/example/demo/Repository/ProductRepository.java
@@ -21,6 +21,9 @@ public interface ProductRepository extends JpaRepository<Products, Integer> {
     @Query("SELECT u FROM Products u WHERE u.productId = :query")
     Products findById(int query);
 
+    @Query("SELECT u FROM Products u WHERE u.productName = :query")
+    Products findByName(String query);
+
     int countByProductName(String ProductName);
 }
